+++ src/HOTP.php
@@ -35,8 +35,8 @@ class HOTP extends AbstractOTP implements OTPInterface
      */
     public function calculate(int $counter = 0) : string
     {
-        $digits = $this->getDigits();
-        $hashFunction = $this->getHashFunction();
+        $digits = $this->options["digits"];
+        $hashFunction = $this->options["hash_func"];
         $secret = $this->secret;
 
         $counter = $this->counterToString($counter);
@@ -56,7 +56,7 @@ class HOTP extends AbstractOTP implements OTPInterface
     public function verify(string $otp, int $counter = 0) : bool
     {
         $counter = max(0, $counter);
-        $window = $this->getWindow();
+        $window = $this->options["window"];
 
         $valid = false;
         $offset = null;
@@ -73,64 +73,6 @@ class HOTP extends AbstractOTP implements OTPInterface
     }
 
     /**
-     * Get the number of digits in the one-time password
-     *
-     * @return  integer Returns the number of digits.
-     */
-    public function getDigits() : int
-    {
-        return $this->digits;
-    }
-
-    /**
-     * Set the number of digits in the one-time password
-     *
-     * @param   integer $digits The number of digits.
-     * @return  self    Returns an instance of self for method chaining.
-     * @throws  InvalidArgumentException    Thrown if the requested number of
-     *                  digits is outside of the inclusive range 1-10.
-     */
-    public function setDigits(int $digits) : self
-    {
-        $digits = abs(intval($digits));
-        if ($digits < 1 || $digits > 10) {
-            throw new InvalidArgumentException("Digits must be a number between 1 and 10 inclusive");
-        }
-        $this->digits = $digits;
-
-        return $this;
-    }
-
-    /**
-     * Get the hash function
-     *
-     * @return  string  Returns the hash function.
-     */
-    public function getHashFunction() : string
-    {
-        return $this->hashFunction;
-    }
-
-    /**
-     * Set the hash function
-     *
-     * @param   string  $hashFunction   The hash function.
-     * @return  self    Returns an instance of self for method chaining.
-     * @throws  InvalidArgumentException    Thrown if the supplied hash function
-     *                                      is not supported.
-     */
-    public function setHashFunction(string $hashFunction) : self
-    {
-        $hashFunction = strtolower($hashFunction);
-        if (!in_array($hashFunction, hash_algos())) {
-            throw new InvalidArgumentException("$hashFunction is not a supported hash function");
-        }
-        $this->hashFunction = $hashFunction;
-
-        return $this;
-    }
-
-    /**
      * Get the counter offset value of the last valid counter value
      *
      * Useful to determine how far ahead the client counter is of the server
@@ -143,46 +85,5 @@ class HOTP extends AbstractOTP implements OTPInterface
     public function getLastValidCounterOffset() : ?int
     {
         return $this->lastCounterOffset;
-    }
-
-    /**
-     * Get the window value
-     *
-     * @return  integer Returns the window value.
-     */
-    public function getWindow() : int
-    {
-        return $this->window;
-    }
-
-    /**
-     * Set the window value
-     *
-     * @param   integer $window The window value.
-     * @return  self    Returns an instance of self for method chaining.
-     */
-    public function setWindow(int $window) : self
-    {
-        $window = abs(intval($window));
-        $this->window = $window;
-
-        return $this;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function processOptions(array $options) : void
-    {
-        // Option names taken from Google Authenticator docs for consistency
-        $options = array_merge([
-            "algorithm" => "sha1",
-            "digits" => 6,
-            "window" => 4,
-        ], array_change_key_case($options, CASE_LOWER));
-
-        $this->setDigits($options["digits"]);
-        $this->setHashFunction($options["algorithm"]);
-        $this->setWindow($options["window"]);
     }
 }