+++ CHANGELOG.md
@@ -10,10 +10,12 @@ This file adheres to the format published by [http://keepachangelog.com/](http:/
  - Update minimum PHP version to >=7.1
  - Re-implement PR #9 - Allow tokens to drift away from real time (@pavarnos)
  - Deprecate validate() method in favor of verify()
+ - Class options must now be passed as array on instantiation
 
 ## Removed
  - Dependency on Rych\Random - no longer needed in PHP 7+
  - Seed and Encoder classes - Shared secrets should be simply passed as raw strings. 
+ - Getters/setters for class options
 
 ## [1.1.1] - 2015-06-29